Fix string matching bug in ensure_default_ovs_flows()
Prior to this patch the the code was checking if "in_port=N" was in the
flow string. But, that could lead to unpredicted values as, for example,
"in_port=1" could match with values such as "in_port=10".

This patches fixes this problem by using a regex that ensures that the
next character after the port number is either a comma or end of line.

Bug found while adding unittests to this method. Tests for this will be
added on a following patch.

Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>
Change-Id: I657b04a5e6ab62d3b89ff8ab1c456e2545df1731