Add support for older OVN versions without Chassis_Private
This patch is adding support for older OVN versions that does not
support Chassis_Private. This is the case of the OVN packaged version in
the Ubuntu LTS distribution, currently it's based on OVN 20.03 and it
does not support Chassis_Private.

Since Ubuntu LTS is used in the OpenStack gate we need to support it for
now. Once OVN package is updated to a later verison we can remove
support for older OVN versions.

Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>
Change-Id: I6faa2b8004aa042f4ff664944d86fdead7259bf5