Account for DVR when exposing FIPs
Currently the agent doesn't take into consideration the DVR configuration
for OpenStack deployments. The nat_addresses column for the SB
Port_Binding table includes information of which chassis should claim
the port but the code was ignoring this.

This patch will account for it and expose the FIPs only in the chassis
that will process the FIP (compute nodes in DVR and gateway nodes for
non DVR configurations).

Story: 2010035
Task: 45341

Signed-off-by: Daniel Alvarez Sanchez <dalvarez@redhat.com>
Change-Id: Ib8b49fcc1496e5c3efa9efa1204762be00cdc419