Ensure ovn-lb on provider networks only get exposed in one chassis
There was a problem where ovn-lb on provider networks were being
exposed in any chassis having a cr-lrp connected to the provider
network where the ovn-lb is, breaking the bgp connectivity.

This patch ensures only the node with the cr-lrp whose associated
subnets (logical switches) are hosting the ovn-lb members are added.

It also ensure reaction to set and unset of routers gateway ports
into the provider networks, with regards to load balancer VIP
BGP management.

Change-Id: I155288b7d07ca72d33cb7859ad32d8f27464b23d