Ensure ovn-lb VIPs on tenant networks are exposed
When the expose_tenant_network flag is enabled, the ovn-lb VIP
should also be exposed. This patch ensures the VIP port create/delete
event is detected

The ovn-lb VIP port is a bit special and it gets created upon
loadbalancer creation. The differentiate information for it is:
- It has no mac
- It has no chassis
- It is of type ""
- It is not up

When one port with those features is being created/deleted, it is
checked if the neutron:cidr information is at the external_ids. If
that so, that information is used to expose the relevant IP, only on the
node where the ovn gateway chassis for the VIP subnet is located.

Change-Id: I4015448951e06a112e4fcf7b06f1efd4d815d77e