Add support to only expose tenant IPv6 GUA IPs
A new flag is added so that we can either expose all the
tenant network IPs (expose_tenant_networks) or just the
IPv6 GUA ones (expose_ipv6_gua_tenant_networks).

If the expose_ipv6_gua_tenant_network flag is enabled, before
exposing the tenant IPs it is checked if they are IPv6 and
of global type, and only in that case they will be advertized.

Story 2010374
Task 46609

Change-Id: Ifa1e25aabb3f6a3fc5f1e5ba5a32bfedadf6765a