Ensure FIP to VIP association is not lost on re-sync
When the VM that has a FIP associated is not using dhcp to get
its IP, the mac entry on the port_binding table is set to
['unknown']. This made the resync to fail on the mac checking
and therefore remove the exposed FIP routes.

This patch checks if the macs is unknown and it is a IP on the
tenant networks. If that is the case we don't check the format
of the mac and call expose_ip with empty IPs, as the IP to expose
in this case is the FIP that will be obtained at the expose_ip
method anyway.

Change-Id: Icdfc1cdc8b759181de3509cd5ddda330270fa6dc