Ensure tenant IPs are exposed when using configdrive
When using configdrive instead of dhcp the port_binding entry for
the tenant IP has ['unknown'] at the mac entry instead of the mac
and IPs of the VM. There is information added about the actual
IPs on the external_ids (neutron:cidrs) that could be used in
that case. This patch ensures this information is processed
if there is an unknown mac entry.

Change-Id: I59dab6fa7e1ec64c86340861e2811c3cdcedfa96