Avoid race condition at _process_lrp_port
There is a chance that a lrp port got just created when
we discover it on a re-sync action and did not yet have the
information about the peer. In that case the port should be skip
until the information is ready. Another option was that the lrp
port was being deleted. In that case it is also the right action
to skip it.

Change-Id: Iddc30a3bfc910fa3e2e390e54015c82f13f3fdfa