Avoid race between withdraw_subnet and cr-lrp association
There is a chance that at the same time of processing a
withdraw_subnet event, its associated cr-lrp got moved to the
local chassis. This will lead to KeyError exception as the
withdraw process moves forward but the event to associate the
cr-lrp to the local chassis has not yet being processed.

In this case there is no need to process the withdraw_subnet
event in the local chassis as it has not yet been exposed. This
patch is adding protection for this case.

Change-Id: I64bd6e99cd6b38247b4a77c8a61860d51c6d116c