Avoid race condition on is_provider_network checkings
There is an option for race conditions on the
"is_provider_network" checkings when doing expose/withdraw actions.
This happens when by the time the action checking if the
port associated datapath is a provider network, the network
associated to the datapath has been deleted, throwing an exception
at ovsdbapp level -- ValueError

This case needs to be handled so that if it happens on a expose
action, the processing is not continued (there is no need to expose
an IP on a subnet that is removed, as the port will be removed too).
If by contract it happens on a withdraw, there are some cases where
it can be dismiss too, as other associated event will take care of
them (such as the case for OVN_LBs or for remote IPs), but needs to be
handled in others (such as IPs on the provider network). This
patch is adding this logic to the driver.

Change-Id: Ia5d832499d70ebae2aa413b2028232d745d9131f