Expose ovn lb on provider network if tenant network is disabled
The logic to expose the ovn lb on the provider network was
relying on the information about the tenant networks connected to
the cr-lrp, that is only filled in when tenant networks are also
exposed. This patch is changing that so that this loadbalancers
are exposed even if the tenant networks are not.

Change-Id: I297af3b87b9683724977c279d0f5f5157abe2672