Ensure proper handling of missing bridge_device at wiring
We need information about the bridge_device when wiring/unwiring
the exposed IPs. If this is not provided, the wiring/unwiring functions
need to return False and the driver need to ensure it is considered,
with proper logging of the failure at exposing/withdrawing.

The missing bridge device may happen if the provider network
is deleted in between or if it was created and not yet present
due to missing sync action

Change-Id: I34c9840b1ac3073c5d50c1c6e5b2d488be9ceca9