Do not process external network that are not provider
It seems it is posible to have external networks (with cr-lrp
associated to them) that are not provider networks (e.g., vxlan
type). In this case we are opting for not exposing through BGP
their IPs as there is no physical infra handling it. This patch
makes sure those networks are not considered when processing the
events associated to them.

Change-Id: Ib436894e2337248b4e37039a1d6690fef406753b