Ensure no wiring is added for OVN exposing method
This patch adds initial support for the OVN exposing method, where
there is no need to write stuff once VMs are created in a host, but
just to advertize them.

Note as only underlay exposing method is currently supported for
the bgp drivers, the bgp utils for wiring/unwiring only consider
these 2 methods.

Change-Id: Id0d18b70f097981289851fd65a02dcaba73a7467