Add vrf_leak call to sync
If the FRR service is restarted, its configuration will not include the
bgp-vrf section that is necessary to advertise the IPs that the OVN BGP
agent exposes (unless the agent is restarted later).
With this patch, the bgp-vrf part is added to the FRR configuration when
the sync is executed periodically.

Change-Id: I54e29a6412fe09e727d8a373fccd085e8d25c941