Add initial support for local OVN cluster instead of kernel-networking
This patch adds support for configuring a local (in-node) OVN cluster
to do the actions that were being done by the kernel networking.
This way, things like ovs-dpdk or HWOL could be supported.

This also includes adaptation to the tox.ini coverage so that
test files are not considered. This was making coveraged artificially
higher that it was, so this patch is also adjusting the minimun
coverage accordingly.

Missing items to be handled in follow up patch sets:
- Documentation
- Support for IPv6
- Support for BFD
- Support for vlan provider networks

Change-Id: Ieff19d88a22d7267bd58f8d1d4ecbcdc668c6bfb
(cherry picked from commit 938cdc26439843c2caaa44daa3cde5ac60453d93)