Ensure all vlan tags are processed
There may be serveral neutron vlan provider networks associated
to a physical bridge. For example, having ovs bridge-mappings
set to "datacentre:br-ex", and then have several provider networks
of vlan type associated to datacentre.

This was not being properly handled on the sync actions as only
the first vlan tag was being processed. This patch ensure all the
vlan tags associated to a bridge are processed and their respective
vlan devices are created

In addition, it seems the way of estimating the coverage has changed,
as now the value is higher, this patch also bumps the minimal coverage
(setting it to 92%), so that the coverage is not reduced unnoticed

Change-Id: I39168fa633f0a69439980c5a2fd163e389cf11fe