Ensure withdrawn events are only processed in relevant nodes
The withdrawn events are evaluating the chassis changes but also the
status changes. When the status was changing it was not being
validated that the chassis was the local one, therefore executing
the withdrawn events in nodes there the IPs were not exposed. This
patch fixes it

Closes-Bug: #2044981
Change-Id: I86faa32a10e7185eac11671470200468fa7c9dda
(cherry picked from commit c0e21895e530bcafa1025be739a905b6ae23a240)