Avoid race when deleting VM with FIP
There is a chance that when deleting a VM with a FIP, by the time
the FIP is being withdrawn, the VM port has its chassis already
removed, which makes the withdrawn event to do nothing, leaving
the FIP exposed. This patch fixes that by removing the FIP if the
VM port associated has no chassis (in the same way as it was done
if the port was already deleted). This means it will be executed
in some nodes that should which do not have the IP exposed, but
that is preferred than leaving the FIP exposed while it should not.

Closes-Bug: #2044980
Change-Id: I3857ea10d8e313216e8f3e7252a0bc54a25b9efe
(cherry picked from commit 4a02101fcc860df5bcb659085842c8bf589d3c4b)