Remove obtaining event classes from strings, vol 2
The current Ubuntu image uses OVN 22.04 which is now almost 2 years old.
The schema contains both Logical_DP_Group and Chassis_Private so it's
safe to remove the code that assumes the tables might not be available.

This is a followup patch for the change
I05598c357b636d3d5d70d7cbefe6acae9dbf8033 that was only for NB OVN BGP
driver.

Change-Id: I8c61c1cff59ca2836cdf4256c921cfdfb291a36a