Disable exposing remote_ips, when only the lrp prefix is sufficient
This also requires to use redist kernel in FRR, so there is a change
here which allows to define the default redistribute options in the FRR
template.

Since now this method is now available, the separate KERNEL_LEAK template
can be removed, as the only difference was the redist kernel, instead of
redist connected.

Related-Bug: #2017886
Change-Id: I570d8c482f3d17d63d66699e402c84dc61787638
(cherry picked from commit 5da36a2638499d1fcb222385019ae4304b18e218)