Handle trimming of vlan interface namings
At the moment vlan interfaces would be trimmed in an unexpected manner,
where instead of the interface vlan UUID will be trimmed when provider
bridge is long enough.
For example, vlan 1001 on `br-provider` will result in interface
`br-provider.100` which can lead to non-unique names, ie if there's also
vlan 1002 on the same br-provider.

Moreover, current code fails to ensure required sysctl parameters are
present, since interface name is trimmed.

This is basically continuation of the fix originally proposed in [1]

[1] https://review.opendev.org/c/openstack/ovn-bgp-agent/+/909788

Closes-Bug: #2054599
Change-Id: Ie1470138eb9ada627deead84863710f7ea43fd76