Fix placement of lsp when external_ids not in sync
When options.requested-chassis is not in sync with
external_ids.neutron:host_id it would pick both hosts, causing duplicate
announcements from more than 1 host.

This has been fixed in change 910305, but was left unchanged for the
sync method, causing issues when the sync interval was re-evaluating all
lsp's on the node.

The code for determining the chassis of a port has been moved from the
base_watcher to driver_utils so the logic for the event is the same as the
logic when fetching the records from the northbound database.

Related-bug: #2049902
Change-Id: I545d6b41fd308eb56e5295657260718dc14868f7
(cherry picked from commit 1bacff1dff825d267c390752fa21592c2e6f6588)