Make in_port consistent type
The in_port structure is always list but in case there is a failure when
obtaining the list of ports on the bridge, it was set to a set type.
This patch initializes it as a list.

TrivialFix

Change-Id: Iec62246f730b9aa80d4145b357da49612b850243
(cherry picked from commit 46abd9b503e6949258e48aaf23c050d42fc7442c)