Ensure that ARP/NDP is enabled for vlan devices
In case VLAN networks are being used in neutron for providing external
connectivity they will be assinged an IP address as well in order to
get ARP/NDP proxying to work.

With that VRF exposure method was relying on broken behaviour, as it
never needed/expected ARP/NDP to be enabled var a vlan rather then bridge
as well as vlan device to exist at all. So skipping vlan_id for evpn makes
sense in this case, as only offload is important.

Closes-Bug: #2088558
Change-Id: Ibffda84801150a18ee8c441b4396afc900d8bc36