Add a static FDB entry to fabric-facing bridges
Without a static entry in the FDB the MAC address we use both for the
internal port of the bridge and the bgp-router LRP is re-learned using
the normal Ethernet switch FDB learning logic: whenever the guest
traffic is sent out via the bgp-router LR, the port is set to the patch
port leading to br-bgp. Sending traffic from the internal port then
rewrites that back. Until then the traffic destined to the internal
port is being sent to br-bgp which results in flow table misses and
drop flows being cached.

Adding a static entry to the FDB effectively disables learning based
on the traffic coming out of the patch port leading to br-bgp.

Closes-Bug: #2089297
Change-Id: I8563adb49db8fa94c51f9ba47275e8c7812997ea