Ensure that bridge exist and UP before adding tables
At the moment OVN NB driver attempts to add routes for the device
which is in DOWN state, which results in exception:
pyroute2.netlink.exceptions.NetlinkError: (100, 'Network is down')

Such scenario is possible in case of migration from regular L2 OVN to
BGP or corner cases where operators didn't manage to ensure device
to be UP in their network config.

Anyway, it's better to handle such cases and ensure required devices
are in proper state before executing actions towards them.

Change-Id: I43fd10d5ae220450afc584a0f6fd40eb95ec8134