Add option to avoid VRF removal
A usecase posisble when operator wants to have a peering established
per-VRF. This would mean, that they need to enslave a unique
interface inside of the VRF and define peers for the router in FRR.

Current logic with removal of the VRF upon disconnect makes such
usecase impossible, as there are no means inside of the agent
to recover FRR and VRF configuration to intended by operator state.

This serves as motivation to allow operator to disable VRF removal
through configuration option.

Closes-Bug: #2097121
Change-Id: I2155508b167f84289de8f6510dfca1d5fc6a2a74