functional: Use shutil to find privsep-helper executable
Previously the executable was based on the VIRTUAL_ENV variable that is
not set in case we want to test against system packages. Using shutil
makes sure the privsep-helper executable is picked even if the venv is
not active.

Closes-Bug: #2098127

Change-Id: Ifbb607a85bdb855641f338d8d6524cb96e94b78f
Signed-off-by: Jakub Libosvar <libosvar@redhat.com>
(cherry picked from commit 7c28babdf742deba0ad4ba01734ef1b9c1a6225e)