Fix arp_proxy LSP option formatting
When multiple provider CIDRs are specified ovn-bgp-agent incorrectly formats
the list of CIDRs passed into the arp_proxy option of the LSP. As a result,
logical flows in SB DB are not set up properly and loss of connectivity
occurs due to the lack of ARP responses.

The format OVN expects is instead with space-separated CIDRs:

https://github.com/ovn-org/ovn/blob/v25.03.0/lib/ovn-util.c#L215-L230
```
/* Extracts the mac, IPv4 and IPv6 addresses from * 'address' which
 * should be of the format "MAC [IP1 IP2 ..] .." where IPn should be a
 * valid IPv4 or IPv6 address and stores them in the 'ipv4_addrs' and
```

https://man7.org/linux/man-pages/man5/ovn-nb.5.html (see "options : arp_proxy")

The MAC in front of CIDRs or IPs is optional. So the correct formatting should
be like this: "192.0.4.0/24 192.0.2.0/24"

Closes-Bug: #2104930
Change-Id: Idcdcd5e4ba876181364c4f0b15072518fe758bc1